move some app starts components around
