+++ static/app/views/performance/mobile/appStarts/components/charts/averageComparisonChart.tsx
